+++ tests/integ/modin/resample/test_resample.py
@@ -9,7 +9,6 @@ import pandas as native_pd
 import pytest
 
 from snowflake.snowpark.modin.plugin._internal.resample_utils import (
-    IMPLEMENTED_AGG_METHODS,
     IMPLEMENTED_DATEOFFSET_STRINGS,
 )
 from tests.integ.modin.utils import (
@@ -19,7 +18,24 @@ from tests.integ.modin.utils import (
 )
 from tests.integ.utils.sql_counter import SqlCounter, sql_count_checker
 
-agg_func = pytest.mark.parametrize("agg_func", IMPLEMENTED_AGG_METHODS)
+agg_func = pytest.mark.parametrize(
+    "agg_func",
+    [
+        "max",
+        "min",
+        "mean",
+        "median",
+        "sum",
+        "std",
+        "var",
+        "count",
+        "size",
+        "first",
+        "last",
+        "quantile",
+        "nunique",
+    ],
+)
 freq = pytest.mark.parametrize("freq", IMPLEMENTED_DATEOFFSET_STRINGS)
 interval = pytest.mark.parametrize("interval", [1, 2, 3, 5, 15])